+++ src/client/src/pages/detail/form/log/logRunModal.tsx
@@ -130,7 +130,11 @@ export const LogRunModal: FC<LogRunModalProps> = ({ logRun, updateLogRun, runs }
     [formMethods],
   );
 
-  const closeDialog = async () => {
+  const cancelDialog = () => {
+    updateLogRun(logRun as LogRun, false);
+  };
+
+  const applyDialog = async () => {
     const isValid = await formMethods.trigger();
     if (!isDirty || isValid) {
       const values = getValues();
@@ -213,8 +217,9 @@ export const LogRunModal: FC<LogRunModalProps> = ({ logRun, updateLogRun, runs }
     <FormDialog
       open={true}
       title={logRun.id === 0 ? t("newLogRun") : (logRun.runNumber ?? "-")}
-      onClose={closeDialog}
-      isCloseDisabled={!formState.isValid && Object.keys(formState.errors).length > 0}>
+      onClose={cancelDialog}
+      onApply={applyDialog}
+      isApplyDisabled={!formState.isValid && Object.keys(formState.errors).length > 0}>
       <FormProvider {...formMethods}>
         <BoreholesCard data-cy="logRun-general" title={t("generalInformation")}>
           <FormContainer>